:sparkles: feat:enemyとの与＆被damage、Healthバー関連追加、各実装同士のマージ(playerHP, slash, swapEffect)

+++ src/Assets/Scripts/Feature/Model/PlayerModel.cs
@@ -201,9 +201,9 @@ namespace Feature.Model
             position.Value = pos;
         }
 
-        public void TakeDamage(int damage)
+        public void TakeDamage(uint damage)
         {
-            health.Value = Mathf.Max(health.Value - damage, 0);
+            health.Value = Mathf.Max((int)(health.Value - damage), 0);
         }
     }
 }
